requirements.txt
Django==4.2.7
django-modeltranslation==0.18.11
django-crispy-forms==2.0
crispy-tailwind==0.5.0
Pillow==10.1.0
python-dotenv==1.0.0